6. Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -12,11 +12,11 @@ int main(void)
 	long long int x;
 	float y;
 
-	printf("Size of char: %zd byte(s)\n", sizeof(u));
-	printf("Size of int: %zd byte(s)\n", sizeof(v));
-	printf("Size of long int: %zd byte(s)\n", sizeof(w));
-	printf("Size of long long int: %zd byte(s)\n", sizeof(x));
-	printf("Size of float: %zd byte(s)\n", sizeof(y));
+	printf("Size of char: %lu byte(s)\n", (unsigned long)sizeof(u));
+	printf("Size of int: %lu byte(s)\n", (unsigned long)sizeof(v));
+	printf("Size of long int: %lu byte(s)\n", (unsigned long)sizeof(w));
+	printf("Size of long long int: %lu byte(s)\n", (unsigned long)sizeof(x));
+	printf("Size of float: %lu byte(s)\n", (unsigned long)sizeof(y));
 
 #if defined(__x86_64__) || defined(__ppc64__)
 	/* Output for 64-bit systems */
